[REFACTORING]  Improving PyGrid's code quality (#535)
* Rename storage -> models

- Rename module (storage -> models)
- Split daatabase models into different source files
- Move db singleton instance to models module

* Update models module

* Update CI test_database

* Fix host_federated_training test

* Update import references

* REFACTORY fl_controller
- Move complete_cycle, average_plan_diffs to cycle_manager
- Fix worker.is_eligible method
- Update report event

* Update models module

* Update Grid Architecture

- Split database models into different modules

* Fix small bugs

* Fix flake issues

* Update imports

* Create routes module

Split route source file into different source files.
- federated
- network
- general

* Update federated.py

* Rename module

- node_routing -> network

* Update routes.federated.py
- Replace mockup vars to database commands described by @Prtfw
