--- app/main/auth/__init__.py
@@ -1,3 +1,0 @@
-from .controller import WorkerController
-
-workers = WorkerController()
