+++ app/main/controller/__init__.py
@@ -1,0 +1,3 @@
+from .fl_controller import FLController
+
+processes = FLController()
