+++ app/main/cycles/__init__.py
@@ -1,0 +1,3 @@
+from .cycle_manager import CycleManager
+
+cycle_manager = CycleManager()
