+++ app/main/models/__init__.py
@@ -1,0 +1,3 @@
+from .model_manager import ModelManager
+
+model_manager = ModelManager()
