+++ app/main/network/__init__.py
@@ -1,0 +1,3 @@
+from .network_manager import NetworkManager
+
+network_manager = NetworkManager()
