+++ app/main/network/nodes.py
@@ -1,0 +1,18 @@
+from .. import db
+
+
+class GridNodes(db.Model):
+    """ Grid Nodes table that represents connected grid nodes.
+    
+        Columns:
+            id (primary_key) : node id, used to recover stored grid nodes (UNIQUE).
+            address: Address of grid node.
+    """
+
+    __tablename__ = "__gridnode__"
+
+    id = db.Column(db.String(), primary_key=True)
+    address = db.Column(db.String())
+
+    def __str__(self):
+        return f"< Grid Node {self.id} : {self.address}>"
