+++ app/main/routes/__init__.py
@@ -1,0 +1,3 @@
+from .federated import *
+from .network import *
+from .general import *
