+++ app/main/syft_assets/__init__.py
@@ -1,0 +1,5 @@
+from .plan_manager import PlanManager
+from .protocol_manager import ProtocolManager
+
+plans = PlanManager()
+protocols = ProtocolManager()
