+++ app/main/workers/__init__.py
@@ -1,0 +1,3 @@
+from .worker_manager import WorkerManager
+
+worker_manager = WorkerManager()
