+++ test/test_database.py
@@ -11,7 +11,15 @@ sys.path.append(".")
 
 
 from flask_sqlalchemy import SQLAlchemy
-from app.main.storage import models
+from app.main.models.ai_model import Model, ModelCheckPoint
+from app.main.syft_assets.protocol import Protocol
+from app.main.syft_assets.plan import Plan
+from app.main.workers.worker import Worker
+from app.main.cycles.worker_cycle import WorkerCycle
+from app.main.processes.fl_process import FLProcess
+from app.main.cycles.cycle import Cycle
+from app.main.processes.config import Config
+from app.main import db
 
 app = Flask(__name__)
 
@@ -21,13 +29,13 @@ BIG_INT = 2 ** 32
 class TestDatabase(unittest.TestCase):
     def setUp(self):
         app.config["SQLALCHEMY_DATABASE_URI"] = "sqlite:///test.db"
-        self.db = models.db
+        self.db = db
         self.db.init_app(app)
         app.app_context().push()
         self.db.create_all()
 
     def testCreatePlan(self):
-        my_plan = models.Plan(
+        my_plan = Plan(
             id=randint(0, BIG_INT),
             value="list of plan values".encode("utf-8"),
             value_ts="torchscript(plan)".encode("utf-8"),
@@ -36,7 +44,7 @@ class TestDatabase(unittest.TestCase):
         self.db.session.commit()
 
     def testCreateProtocol(self):
-        my_protocol = models.Protocol(
+        my_protocol = Protocol(
             id=randint(0, BIG_INT),
             value="list of protocol values".encode("utf-8"),
             value_ts="torchscript(protocol)".encode("utf-8"),
@@ -63,16 +71,16 @@ class TestDatabase(unittest.TestCase):
             "minimum_download_speed": 4000,  # 4 mbps
         }
 
-        my_server_config = models.Config(id=randint(0, BIG_INT), config=server_config)
+        my_server_config = Config(id=randint(0, BIG_INT), config=server_config)
 
-        my_client_config = models.Config(id=randint(0, BIG_INT), config=client_config)
+        my_client_config = Config(id=randint(0, BIG_INT), config=client_config)
 
         self.db.session.add(my_server_config)
         self.db.session.add(my_client_config)
         self.db.session.commit()
 
     def testCreateWorker(self):
-        worker = models.Worker(
+        worker = Worker(
             id=randint(0, BIG_INT),
             format_preference="list",
             ping=randint(0, 100),
@@ -84,7 +92,7 @@ class TestDatabase(unittest.TestCase):
         self.db.session.commit()
 
     def testCreateCycle(self):
-        new_cycle = models.Cycle(
+        new_cycle = Cycle(
             id=randint(0, BIG_INT),
             start=datetime(2019, 2, 21, 7, 29, 32, 45),
             sequence=randint(0, 100),
@@ -95,21 +103,21 @@ class TestDatabase(unittest.TestCase):
         self.db.session.commit()
 
     def testCreateModel(self):
-        new_model = models.Model(version="0.0.1")
+        new_model = Model(version="0.0.1")
 
         self.db.session.add(new_model)
         self.db.session.commit()
 
     def testCreateFLProcess(self):
-        new_fl_process = models.FLProcess(id=randint(0, BIG_INT))
+        new_fl_process = FLProcess(id=randint(0, BIG_INT))
 
         self.db.session.add(new_fl_process)
 
-        new_model = models.Model(version="0.0.1", flprocess=new_fl_process)
+        new_model = Model(version="0.0.1", flprocess=new_fl_process)
 
         self.db.session.add(new_model)
 
-        avg_plan = models.Plan(
+        avg_plan = Plan(
             id=randint(0, BIG_INT),
             value="list of plan values".encode("utf-8"),
             value_ts="torchscript(plan)".encode("utf-8"),
@@ -118,7 +126,7 @@ class TestDatabase(unittest.TestCase):
 
         self.db.session.add(avg_plan)
 
-        training_plan = models.Plan(
+        training_plan = Plan(
             id=randint(0, BIG_INT),
             value="list of plan values".encode("utf-8"),
             value_ts="torchscript(plan)".encode("utf-8"),
@@ -127,7 +135,7 @@ class TestDatabase(unittest.TestCase):
 
         self.db.session.add(training_plan)
 
-        validation_plan = models.Plan(
+        validation_plan = Plan(
             id=randint(0, BIG_INT),
             value="list of plan values".encode("utf-8"),
             value_ts="torchscript(plan)".encode("utf-8"),
@@ -136,7 +144,7 @@ class TestDatabase(unittest.TestCase):
 
         self.db.session.add(validation_plan)
 
-        protocol_1 = models.Protocol(
+        protocol_1 = Protocol(
             id=randint(0, BIG_INT),
             value="list of protocol values".encode("utf-8"),
             value_ts="torchscript(protocol)".encode("utf-8"),
@@ -145,7 +153,7 @@ class TestDatabase(unittest.TestCase):
 
         self.db.session.add(protocol_1)
 
-        protocol_2 = models.Protocol(
+        protocol_2 = Protocol(
             id=randint(0, BIG_INT),
             value="list of protocol values".encode("utf-8"),
             value_ts="torchscript(protocol)".encode("utf-8"),
@@ -172,13 +180,13 @@ class TestDatabase(unittest.TestCase):
             "minimum_download_speed": 4000,  # 4 mbps
         }
 
-        server_config = models.Config(
+        server_config = Config(
             id=randint(0, BIG_INT),
             config=server_config,
             server_flprocess_config=new_fl_process,
         )
 
-        client_config = models.Config(
+        client_config = Config(
             id=randint(0, BIG_INT),
             config=client_config,
             client_flprocess_config=new_fl_process,
@@ -187,7 +195,7 @@ class TestDatabase(unittest.TestCase):
         self.db.session.add(client_config)
         self.db.session.add(server_config)
 
-        cycle_1 = models.Cycle(
+        cycle_1 = Cycle(
             id=randint(0, BIG_INT),
             start=datetime(2019, 2, 21, 7, 29, 32, 45),
             sequence=randint(0, 100),
@@ -195,7 +203,7 @@ class TestDatabase(unittest.TestCase):
             cycle_flprocess=new_fl_process,
         )
 
-        cycle_2 = models.Cycle(
+        cycle_2 = Cycle(
             id=randint(0, BIG_INT),
             start=datetime(2019, 2, 27, 15, 19, 22),
             sequence=randint(0, 100),
@@ -208,15 +216,15 @@ class TestDatabase(unittest.TestCase):
         self.db.session.commit()
 
     def testWorkerCycle(self):
-        new_fl_process = models.FLProcess(id=randint(0, BIG_INT))
+        new_fl_process = FLProcess(id=randint(0, BIG_INT))
 
         self.db.session.add(new_fl_process)
 
-        new_model = models.Model(version="0.0.1", flprocess=new_fl_process)
+        new_model = Model(version="0.0.1", flprocess=new_fl_process)
 
         self.db.session.add(new_model)
 
-        avg_plan = models.Plan(
+        avg_plan = Plan(
             id=randint(0, BIG_INT),
             value="list of plan values".encode("utf-8"),
             value_ts="torchscript(plan)".encode("utf-8"),
@@ -225,7 +233,7 @@ class TestDatabase(unittest.TestCase):
 
         self.db.session.add(avg_plan)
 
-        training_plan = models.Plan(
+        training_plan = Plan(
             id=randint(0, BIG_INT),
             value="list of plan values".encode("utf-8"),
             value_ts="torchscript(plan)".encode("utf-8"),
@@ -234,7 +242,7 @@ class TestDatabase(unittest.TestCase):
 
         self.db.session.add(training_plan)
 
-        validation_plan = models.Plan(
+        validation_plan = Plan(
             id=randint(0, BIG_INT),
             value="list of plan values".encode("utf-8"),
             value_ts="torchscript(plan)".encode("utf-8"),
@@ -243,7 +251,7 @@ class TestDatabase(unittest.TestCase):
 
         self.db.session.add(validation_plan)
 
-        protocol = models.Protocol(
+        protocol = Protocol(
             id=randint(0, BIG_INT),
             value="list of protocol values".encode("utf-8"),
             value_ts="torchscript(protocol)".encode("utf-8"),
@@ -270,13 +278,13 @@ class TestDatabase(unittest.TestCase):
             "minimum_download_speed": 4000,  # 4 mbps
         }
 
-        server_config = models.Config(
+        server_config = Config(
             id=randint(0, BIG_INT),
             config=server_config,
             server_flprocess_config=new_fl_process,
         )
 
-        client_config = models.Config(
+        client_config = Config(
             id=randint(0, BIG_INT),
             config=client_config,
             client_flprocess_config=new_fl_process,
@@ -285,7 +293,7 @@ class TestDatabase(unittest.TestCase):
         self.db.session.add(client_config)
         self.db.session.add(server_config)
 
-        cycle = models.Cycle(
+        cycle = Cycle(
             id=randint(0, BIG_INT),
             start=datetime(2019, 2, 21, 7, 29, 32, 45),
             sequence=randint(0, 100),
@@ -295,7 +303,7 @@ class TestDatabase(unittest.TestCase):
 
         self.db.session.add(cycle)
 
-        worker = models.Worker(
+        worker = Worker(
             id=randint(0, BIG_INT),
             format_preference="list",
             ping=randint(0, 100),
@@ -305,7 +313,7 @@ class TestDatabase(unittest.TestCase):
 
         self.db.session.add(worker)
 
-        worker_cycle = models.WorkerCycle(
+        worker_cycle = WorkerCycle(
             id=randint(0, BIG_INT),
             request_key="long_hashcode_here",
             worker=worker,
